fix: fix bug with toast not appearing
